move observation filter to default
