--- dubbo-metrics/dubbo-metrics-api/src/main/resources/META-INF/dubbo/internal/org.apache.dubbo.rpc.Filter
@@ -1,1 +1,0 @@
-observationreceiver=org.apache.dubbo.metrics.filter.observation.ObservationReceiverFilter
