+++ dubbo-metrics/dubbo-metrics-default/src/main/java/org/apache/dubbo/metrics/observation/ObservationSenderFilter.java
@@ -14,35 +14,38 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.dubbo.metrics.filter.observation;
+package org.apache.dubbo.metrics.observation;
 
 import io.micrometer.observation.Observation;
 import io.micrometer.observation.ObservationRegistry;
 import org.apache.dubbo.common.extension.Activate;
+import org.apache.dubbo.metrics.filter.observation.DefaultDubboClientObservationConvention;
+import org.apache.dubbo.metrics.filter.observation.DubboObservation;
 import org.apache.dubbo.rpc.BaseFilter;
 import org.apache.dubbo.rpc.Filter;
 import org.apache.dubbo.rpc.Invocation;
 import org.apache.dubbo.rpc.Invoker;
 import org.apache.dubbo.rpc.Result;
 import org.apache.dubbo.rpc.RpcException;
+import org.apache.dubbo.rpc.cluster.filter.ClusterFilter;
 import org.apache.dubbo.rpc.model.ApplicationModel;
 import org.apache.dubbo.rpc.model.ScopeModelAware;
 
-import static org.apache.dubbo.common.constants.CommonConstants.PROVIDER;
+import static org.apache.dubbo.common.constants.CommonConstants.CONSUMER;
 
 /**
- * A {@link Filter} that creates an {@link Observation} around the incoming message.
+ * A {@link Filter} that creates an {@link Observation} around the outgoing message.
  */
-@Activate(group = PROVIDER, order = -1)
-public class ObservationReceiverFilter implements Filter, BaseFilter.Listener, ScopeModelAware {
+@Activate(group = CONSUMER, order = -1)
+public class ObservationSenderFilter implements ClusterFilter, BaseFilter.Listener, ScopeModelAware {
 
     private final ObservationRegistry observationRegistry;
 
-    private final DubboServerObservationConvention serverObservationConvention;
+    private final org.apache.dubbo.metrics.filter.observation.DubboClientObservationConvention clientObservationConvention;
 
-    public ObservationReceiverFilter(ApplicationModel applicationModel) {
+    public ObservationSenderFilter(ApplicationModel applicationModel) {
         observationRegistry = applicationModel.getBeanFactory().getBean(ObservationRegistry.class);
-        serverObservationConvention = applicationModel.getBeanFactory().getBean(DubboServerObservationConvention.class);
+        clientObservationConvention = applicationModel.getBeanFactory().getBean(org.apache.dubbo.metrics.filter.observation.DubboClientObservationConvention.class);
     }
 
     @Override
@@ -50,8 +53,8 @@ public class ObservationReceiverFilter implements Filter, BaseFilter.Listener, S
         if (observationRegistry == null) {
             return invoker.invoke(invocation);
         }
-        DubboServerContext receiverContext = new DubboServerContext(invoker, invocation);
-        Observation observation = DubboObservation.SERVER.observation(this.serverObservationConvention, DefaultDubboServerObservationConvention.INSTANCE, () -> receiverContext, observationRegistry);
+        org.apache.dubbo.metrics.filter.observation.DubboClientContext senderContext = new org.apache.dubbo.metrics.filter.observation.DubboClientContext(invoker, invocation);
+        Observation observation = DubboObservation.CLIENT.observation(this.clientObservationConvention, DefaultDubboClientObservationConvention.INSTANCE, () -> senderContext, observationRegistry);
         invocation.put(Observation.class, observation.start());
         return observation.scoped(() -> invoker.invoke(invocation));
     }
